refactor: move MFN logic out of AsmDB index file
